Replace [Button] to [SimpleButton] and rename it to [Button].
