+++ src/components/shared/__tests__/Button.test.tsx
@@ -1,44 +1,56 @@
+import 'react-native';
+
 import * as React from 'react';
 
-import Button from '../Button';
-import { Text } from 'react-native';
+import Button, { THEME } from '../Button';
+import {
+  RenderResult,
+  act,
+  fireEvent,
+  render,
+} from '@testing-library/react-native';
+
+import { ThemeProvider } from 'styled-components/native';
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
-const component = (props?: any): React.ReactElement => {
-  return <Button {...props} />;
-};
+let props: any;
+let component: React.ReactElement;
+let testingLib: RenderResult;
 
 describe('[Button]', () => {
-  let rendered: renderer.ReactTestRenderer;
+  let cnt = 1;
 
-  it('should render without crashing', () => {
-    rendered = renderer.create(component());
-    expect(rendered.toJSON()).toMatchSnapshot();
-    expect(rendered.toJSON()).toBeTruthy();
+  beforeEach(() => {
+    props = {
+      onClick: (): number => cnt++,
+      testID: 'btn',
+    };
+    component = (
+      <ThemeProvider theme={THEME.LIGHT}>
+        <Button {...props} />
+      </ThemeProvider>
+    );
   });
 
-  describe('[Button] Interaction', () => {
-    let root: renderer.ReactTestInstance;
-    let cnt = 1;
-    it('simulate onPress', () => {
-      rendered = renderer.create(
-        component({
-          onClick: () => cnt++,
-        }),
-      );
-      root = rendered.root;
+  it('renders without crashing', () => {
+    const rendered = renderer.create(component).toJSON();
+    expect(rendered).toMatchSnapshot();
+    expect(rendered).toBeTruthy();
+  });
 
-      root.findByType(Button).props.onClick();
-      expect(cnt).toBe(2);
+  describe('interactions', () => {
+    beforeEach(() => {
+      testingLib = render(component);
     });
 
-    it('renders disabled', () => {
-      rendered = renderer.create(component({ isDisabled: true }));
-      root = rendered.root;
-
-      const texts = root.findAllByType(Text);
-      expect(texts).toHaveLength(1);
+    it('should simulate onClick', () => {
+      const btn = testingLib.queryByTestId('btn');
+      act(() => {
+        fireEvent.press(btn);
+        fireEvent.press(btn);
+      });
+      expect(cnt).toBe(3);
     });
   });
 });
